Cinder liberty-2 milestone (7.0.0.0b2)

Conflicts:
	cinder/volume/drivers/zfssa/zfssaiscsi.py

+++ cinder/volume/drivers/san/hp/hp_3par_common.py
@@ -52,6 +52,8 @@ if hp3parclient:
 
 from oslo_config import cfg
 from oslo_log import log as logging
+from oslo_log import versionutils
+from oslo_service import loopingcall
 from oslo_utils import excutils
 from oslo_utils import units
 
@@ -59,8 +61,6 @@ from cinder import context
 from cinder import exception
 from cinder import flow_utils
 from cinder.i18n import _, _LE, _LI, _LW
-from cinder.openstack.common import loopingcall
-from cinder.openstack.common import versionutils
 from cinder.volume import qos_specs
 from cinder.volume import utils as volume_utils
 from cinder.volume import volume_types
@@ -178,10 +178,12 @@ class HP3PARCommon(object):
         2.0.43 - Report the capability of supporting multiattach
         2.0.44 - Update help strings to reduce the 3PAR user role requirements
         2.0.45 - Python 3 fixes
+        2.0.46 - Improved VLUN creation and deletion logic. #1469816
+        2.0.47 - Changed initialize_connection to use getHostVLUNs. #1475064
 
     """
 
-    VERSION = "2.0.45"
+    VERSION = "2.0.47"
 
     stats = {}
 
@@ -794,24 +796,57 @@ class HP3PARCommon(object):
         volume_name = self._get_3par_vol_name(volume['id'])
         vluns = self.client.getHostVLUNs(hostname)
 
+        # Find all the VLUNs associated with the volume. The VLUNs will then
+        # be split into groups based on the active status of the VLUN. If there
+        # are active VLUNs detected a delete will be attempted on them. If
+        # there are no active VLUNs but there are inactive VLUNs, then the
+        # inactive VLUNs will be deleted. The inactive VLUNs are the templates
+        # on the 3PAR backend.
+        active_volume_vluns = []
+        inactive_volume_vluns = []
+        volume_vluns = []
+
         for vlun in vluns:
             if volume_name in vlun['volumeName']:
-                break
-        else:
-            LOG.info(_LI("3PAR vlun for volume %(name)s not found on host "
-                         "%(host)s"), {'name': volume_name, 'host': hostname})
+                if vlun['active']:
+                    active_volume_vluns.append(vlun)
+                else:
+                    inactive_volume_vluns.append(vlun)
+        if active_volume_vluns:
+            volume_vluns = active_volume_vluns
+        elif inactive_volume_vluns:
+            volume_vluns = inactive_volume_vluns
+
+        if not volume_vluns:
+            msg = (
+                _LW("3PAR vlun for volume %(name)s not found on "
+                    "host %(host)s"), {'name': volume_name, 'host': hostname})
+            LOG.warning(msg)
             return
 
         # VLUN Type of MATCHED_SET 4 requires the port to be provided
-        if self.VLUN_TYPE_MATCHED_SET == vlun['type']:
-            self.client.deleteVLUN(volume_name, vlun['lun'], hostname,
-                                   vlun['portPos'])
-        else:
-            self.client.deleteVLUN(volume_name, vlun['lun'], hostname)
+        removed_luns = []
+        for vlun in volume_vluns:
+            if self.VLUN_TYPE_MATCHED_SET == vlun['type']:
+                self.client.deleteVLUN(volume_name, vlun['lun'], hostname,
+                                       vlun['portPos'])
+            else:
+                # This is HOST_SEES or a type that is not MATCHED_SET.
+                # By deleting one VLUN, all the others should be deleted, too.
+                if vlun['lun'] not in removed_luns:
+                    self.client.deleteVLUN(volume_name, vlun['lun'], hostname)
+                    removed_luns.append(vlun['lun'])
 
         # Determine if there are other volumes attached to the host.
         # This will determine whether we should try removing host from host set
         # and deleting the host.
+        vluns = []
+        try:
+            vluns = self.client.getHostVLUNs(hostname)
+        except hpexceptions.HTTPNotFound:
+            LOG.debug("All VLUNs removed from host %s", hostname)
+            pass
+
         for vlun in vluns:
             if volume_name not in vlun['volumeName']:
                 # Found another volume
@@ -2054,6 +2089,30 @@ class HP3PARCommon(object):
         old_volume_settings = self.get_volume_settings_from_type(volume, host)
         return self._retype_from_old_to_new(volume, new_type,
                                             old_volume_settings, host)
+
+    def find_existing_vlun(self, volume, host):
+        """Finds an existing VLUN for a volume on a host.
+
+        Returns an existing VLUN's information. If no existing VLUN is found,
+        None is returned.
+
+        :param volume: A dictionary describing a volume.
+        :param host: A dictionary describing a host.
+        """
+        existing_vlun = None
+        try:
+            vol_name = self._get_3par_vol_name(volume['id'])
+            host_vluns = self.client.getHostVLUNs(host['name'])
+
+            # The first existing VLUN found will be returned.
+            for vlun in host_vluns:
+                if vlun['volumeName'] == vol_name:
+                    existing_vlun = vlun
+                    break
+        except hpexceptions.HTTPNotFound:
+            # ignore, no existing VLUNs were found
+            pass
+        return existing_vlun
 
     class TaskWaiter(object):
         """TaskWaiter waits for task to be not active and returns status."""
